score, pins, and more
I added a scoring system. It doesn't work yet but I did get the system to recognize whether a strike is thrown, a spare is thrown, or if the player got nine or less pins. I also changed some of the mass of the pins and slightly increased the amount that pins have to move to be considered knocked over.