coins for bonus
Because i can't figure out how to code the logic for strikes and spares, i added a coin bonus and just added 10 to the score. the coins don't have a function yet, but they are determined by the streak multiplier and if the player threw a strike or a spare. I also added the mechanic where after 10 frames, your score for the game gets added to the coin balanced. I haven't added the mechanic when after 10 frames the game ends.